hoist toast props out of toast component
