+++ libs/ui/src/hooks/useToast.tsx
@@ -32,6 +32,7 @@ export const useToast = () => {
       iconType: 'success',
     });
   };
+
   const warningToast = ({
     title,
     description,
